requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
amqp==5.1.1; python_version >= '3.6'
arabic-reshaper==2.1.3
asgiref==3.5.2; python_version >= '3.7'
asn1crypto==1.5.1
asttokens==2.0.5
backcall==0.2.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15; python_version >= '3.6'
cffi==1.15.1
charset-normalizer==2.1.0; python_version >= '3.6'
click-didyoumean==0.3.0; python_version < '4' and python_full_version >= '3.6.2'
click-plugins==1.1.1
click-repl==0.2.0
click==8.1.3; python_version >= '3.7'
colorama==0.4.5; sys_platform == 'win32'
cryptography==37.0.2; python_version >= '3.6'
cssselect2==0.6.0; python_version >= '3.7'
decorator==5.1.1; python_version >= '3.5'
defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
dj-rest-auth==2.2.4
django-allauth==0.51.0
django-colorfield==0.7.1
django-phonenumber-field[phonenumbers]==6.3.0
django-truncate==0.1
django==4.0.5
djangorestframework-simplejwt==5.2.0
djangorestframework==3.13.1
executing==0.8.3
future==0.18.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
html5lib==1.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
idna==3.3; python_version >= '3.5'
ipython==8.4.0
jedi==0.18.1; python_version >= '3.6'
kombu==5.2.4; python_version >= '3.7'
lxml==4.9.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
matplotlib-inline==0.1.3; python_version >= '3.5'
oauthlib==3.2.0; python_version >= '3.6'
oscrypto==1.3.0
parso==0.8.3; python_version >= '3.6'
phonenumbers==8.12.51
pickleshare==0.7.5
pillow==9.2.0
prompt-toolkit==3.0.30; python_full_version >= '3.6.2'
pure-eval==0.2.2
pycparser==2.21
pygments==2.12.0; python_version >= '3.6'
pyhanko-certvalidator==0.19.5
pyhanko==0.13.2
pyjwt[crypto]==2.4.0; python_version >= '3.6'
pypdf3==1.0.6
python-bidi==0.4.2
python-http-client==3.3.7; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python3-openid==3.2.0
pytz-deprecation-shim==0.1.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pytz==2022.1
pyyaml==6.0; python_version >= '3.6'
qrcode==7.3.1; python_version >= '3.6'
reportlab==3.6.10; python_version >= '3.7' and python_version < '4'
requests-oauthlib==1.3.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
requests==2.28.1; python_version >= '3.7' and python_version < '4'
sendgrid==6.9.7
setuptools==63.1.0; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.2; python_version >= '3.5'
stack-data==0.3.0
starkbank-ecdsa==2.0.3
svglib==1.3.0; python_version >= '3.7'
tinycss2==1.1.1; python_version >= '3.6'
tqdm==4.64.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
traitlets==5.3.0; python_version >= '3.7'
tzdata==2022.1; sys_platform == 'win32'
tzlocal==4.2; python_version >= '3.6'
uritools==4.0.0; python_version ~= '3.7'
urllib3==1.26.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
vine==5.0.0; python_version >= '3.6'
wcwidth==0.2.5
webencodings==0.5.1
xhtml2pdf==0.2.8
